Add files via upload
Recommendation system that recommends movies that are similar to a particular movie. To achieve this, we compute pairwise cosine similarity scores for all movies based on their plot descriptions and recommend movies based on that similarity score threshold.
The plot description is available as the overview feature in  "movies_metadata" dataset.